remove siwe feature gate, fix siwe history issues and rename siwe mod
